Add ranking and voting to each story

+++ db/migrate/20180526194826_add_rank_to_stories.rb
@@ -1,5 +1,5 @@
 class AddRankToStories < ActiveRecord::Migration[5.2]
   def change
-    add_column :stories, :rank, :integer, default: 0
+    add_column :stories, :rank, :float, default: 0
   end
 end
